Handle self-insert with completions
Currently, copilot.el will always clear the suggestion when and character is
typed. this leads to the suggestion often disappearing as soon as it pops up when
typing quickly or when working over a slow connection, even if it's completing
exactly what you are typing.

This change makes it so that if the characters inserted are already the start of
the suggestion, then it will keep the suggestion visible.