support middle-of-the-line completions
GitHub Copilot usually returns no completions if position is not at
the end of the line. However if position is only one character
from the end completions are sometimes returned, but the behavior
is unreliable - sometimes the last character is returned as part of
completion, sometimes not.

This commit attempts to handle this case by removing common suffix
between completion and current line prior to insertion.

I also converted overlay code to use zero-width overlays at all times
- in order to unify the code paths and get rid of special cases for
eof and eol.

For some reason `(put-text-property 0 1 'cursor t)` didn't work as
expected - so I had to replace it with
`(put-text-property 0 1 'cursor 1)`
for now, will investigate further.