Fix handling of completions when they don't go to the end of line
Currently copilot.el assumes that the completion text will go all the way to the
end of the line. But this is not always the case. When this happens it
will lead to incorrectly deleting the text that used to be there. This will lead
to unbalanced parens or quotes.

The fix is to use the API's :end property to determine where the end of the
completed text should go.

Example:
text given to copilot
```python
print("hello |")
```
copilot will return the text `world`.

With the current implementation this will be the result if the user accepts the completion:
```python
print("hello world
```

After this PR, the result would be as expected:
```python
print("hello world")
```