balancer: add a case where end string quote needs to be overwritten
Sometimes a returned completion needs to be have the 'end' incremented
by one so that the end quote is overwritten. This seems to occur when
you are inside a string and start == end and the char after the end
point is a double quote. We attempt to detect these conditions by
seeing if there is an odd number of double quotes in the returned
completion. In the code this is the `infix-string-fixup-needed`
variable.

However, the implement this, the main copilot.el file also needs to be modified.