more lisp balancer fixes (#180)
* balancer: fix misbalance when in a multi-line string

* copilot-balancer: rewrite debug to use plist

This makes it less fragile to ordering of the arguments.

* copilot-balancer: fix not seeing end of string in some cases

Sometimes inside a string there are pair characters, making it think
the string is not terminated. The only way to fix this is to keep
scanning for a string, but we should be able to terminate after the
top level defun.

* copilot-balancer: fix missing dquote search being too eager

Given a long enough top level defun, you will always find a double
quote, so we must be more careful.

One way is to check if the delete text contains a double quote.

The other way is to cap it by the number of meta prefix pair chars.

* copilot-balancer: fix bug checking if end is missing a double quote

* balancer: add a case where end string quote needs to be overwritten

Sometimes a returned completion needs to be have the 'end' incremented
by one so that the end quote is overwritten. This seems to occur when
you are inside a string and start == end and the char after the end
point is a double quote. We attempt to detect these conditions by
seeing if there is an odd number of double quotes in the returned
completion. In the code this is the `infix-string-fixup-needed`
variable.

However, the implement this, the main copilot.el file also needs to be modified.

* balancer: infix string check probably does not need to be in a string