fix: use `standard-indent` instead `tab-width` for indentation (#335)
I see the following warnings.

```
Warning: copilot--infer-indentation-offset found no mode-specific indentation offset, using ’tab-width’ instead.  You can suppress this error message by customizing ’copilot-indent-warning-suppress’.
```

Reading the documentation comments for `tab-width` shows that it has no effect if spaces are used for indentation.

```
Documentation
Distance between tab stops (for display of tab characters), in columns.

This controls the width of a TAB character on display.
The value should be a positive integer.
Note that this variable doesn't necessarily affect the size of the
indentation step.  However, if the major mode's indentation facility
inserts one or more TAB characters, this variable will affect the
indentation step as well, even if indent-tabs-mode is non-nil.
```

If tabs are used, the `tab-width` is automatically adjusted, so there seems to be no need to worry about it.

Conversely, `tabSize` is supposed to be the real tab size, not the indentation size, so put the value as it is.